+++ packages/mask/src/extension/background-script/EthereumServices/send.ts
@@ -11,7 +11,7 @@ import {
     EthereumMethodType,
     EthereumRpcType,
     EthereumTransactionConfig,
-    getTokenConstants,
+    ZERO_ADDRESS,
     isEIP1559Supported,
     isSameAddress,
     ProviderType,
@@ -138,8 +138,7 @@ async function handleTransferTransaction(chainId: ChainId, payload: JsonRpcPaylo
     const from = (computedPayload._tx.from as string) ?? ''
     const to = getTo(computedPayload)
 
-    if (!isSameAddress(from, to) && !isSameAddress(to, getTokenConstants(ChainId.Mainnet).ZERO_ADDRESS))
-        await WalletRPC.addAddress(chainId, to)
+    if (!isSameAddress(from, to) && !isSameAddress(to, ZERO_ADDRESS)) await WalletRPC.addAddress(chainId, to)
 }
 
 function handleRecentTransaction(
