+++ packages/web3-constants/evm/token.json
@@ -1,18 +1,4 @@
 {
-    "ZERO_ADDRESS": {
-        "Mainnet": "0x0000000000000000000000000000000000000000",
-        "Ropsten": "0x0000000000000000000000000000000000000000",
-        "Rinkeby": "0x0000000000000000000000000000000000000000",
-        "Kovan": "0x0000000000000000000000000000000000000000",
-        "Gorli": "0x0000000000000000000000000000000000000000",
-        "BSC": "0x0000000000000000000000000000000000000000",
-        "BSCT": "0x0000000000000000000000000000000000000000",
-        "Matic": "0x0000000000000000000000000000000000000000",
-        "Mumbai": "0x0000000000000000000000000000000000000000",
-        "Arbitrum": "0x0000000000000000000000000000000000000000",
-        "Arbitrum_Rinkeby": "0x0000000000000000000000000000000000000000",
-        "xDai": "0x0000000000000000000000000000000000000000"
-    },
     "WNATIVE_ADDRESS": {
         "Mainnet": "0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2",
         "Ropsten": "0xc778417E063141139Fce010982780140Aa0cD5Ab",
