+++ packages/web3-shared/evm/constants/index.ts
@@ -18,6 +18,8 @@ import OpenseaAPI from '@masknet/web3-constants/evm/opensea-api.json'
 import Chain from '@masknet/web3-constants/evm/chain.json'
 import { hookTransform, transform, transformFromJSON } from './utils'
 
+export { ZERO_ADDRESS, FAKE_SIGN_PASSWORD, EthereumNameType } from './specific'
+
 export const getAirdropConstants = transform(Airdrop)
 export const useAirdropConstants = hookTransform(getAirdropConstants)
 
@@ -75,13 +77,3 @@ export const useChainConstants = hookTransform(getChainConstants)
 
 export const getNftRedPacketConstants = transform(NftRedPacket)
 export const useNftRedPacketConstants = hookTransform(getNftRedPacketConstants)
-
-// for estimate gas
-export const FAKE_SIGN_PASSWORD = '0x75466cc969717b172b14253aaeebdc958f2b5037a852c1337650ed4978242dd9'
-
-// for ens address
-export enum EthereumNameType {
-    ENS_TYPE = 'ENS',
-    UNS_TYPE = 'UNS',
-    DEFAULT = 'address',
-}
