+++ packages/web3-shared/evm/constants/specific.ts
@@ -1,0 +1,11 @@
+export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
+
+// for estimate gas
+export const FAKE_SIGN_PASSWORD = '0x75466cc969717b172b14253aaeebdc958f2b5037a852c1337650ed4978242dd9'
+
+// for ens address
+export enum EthereumNameType {
+    ENS_TYPE = 'ENS',
+    UNS_TYPE = 'UNS',
+    DEFAULT = 'address',
+}
